New html for the landing page
This changes the design to more closely resemble the initial design. I am creating a pull request in case this could change something important that was added that I wasn't aware of. Github's 99 file limit wouldn't let me upload the entire directory, so the file paths here will still need to be put in the correct locations on the finalization